Fix reading runtime params in HDF5 dump header
Some run time parameters such as, `tolh`, `C_cour`, `C_force`, etc. were
being read from the HDF5 dump file and set there. When they should be
read from the infile. Now they are read from file but the values are not
used to set their values globally. The values do not need to be read but
this follows the code as written for "regular" Phantom dump files in
`readwrite_dumps.F90`.