Parallel build action (#224)
Type of PR:
other

Description:
The build check action makes each configuration sequentially. Currently there are 93 setups, so this takes about 3 hours to finish. This test slows down development pace because developers must wait a long time for checks on pull requests.

While waiting for build checks on my other PRs to finish, I parallelised the build action 😄

The maximum github action matrix size is 128. The tests cannot be run individually because all 93 tests have to be run for each compiler, >128. buildbot.sh has been modified to run setups in batches. It takes the argument

./buildbot.sh --parallel i j
where j is the number of batches to split the setups into, and the i'th batch is run. For example, --parallel 2 10 will split the setups into 10 batches, and run the 2nd one. To run all the setups, the command must be called j times, for i=1 to i=j. This is split across github runners.

j is set to 32. Using the github runner (for gfortran compiler), the entire build test runs in ~12 minutes, which is ~15x faster. The ifort compiler will be bottlenecked by the number of available custom runners.

A summary job is added to the end of all workflows with a matrix so that the branch protection requirement does not need to be changed when the matrix changes. The requirements build, mpi, test will now cover all tests. Resolves #226

Testing:
build test continues to run all setups

Did you run the bots? no - not required