Fix sink potential check (#234)
Type of PR: 
Bug fix

Description:
Required for #217, discovered when `mpirun -n 4` was added

In ptmass.F90, during sink creation, whether the particle is at the potential minimum is checked:
```
       ! CHECK 7: particle i is at minimum in potential
       if (poteni > potenj_min) then
          ifail = inosink_poten
          ifail_array(inosink_poten) = 1
       endif
```

This check fails because an identical particle distribution gets created on every task, resulting in close comparisons when determining which particle to create a sink at.

This is fixed by only creating the particles on the master task, and allowing derivs to rebalance them.

`itest` needs to be calculated after calling derivs or else it will be incorrect. The `xcofm` calculation has been removed because the result is not used anywhere.

Testing:
```
make MPI=yes SETUP=testgrav phantomtest
mpirun -n 4 bin/phantomtest ptmass
```

Did you run the bots? yes, pre-commit