remove accradius1_hard hack (#245)
This hard accretion radius hack was introduced in https://github.com/danieljprice/phantom/commit/e54d7350c10c88c7e957227e995b2ff50cb3e3fe to test a scenario in response to a referee's request. It was merged into master, but is not useful in general. It also causes issues elsewhere because boundary particles can be created with no `massoftype set`. @dliptai and @danieljprice have agreed that it should be removed.